Frontend events and news controllers fixed, admin controllers refactored

+++ application/modules/admin/models/EventsMapper.php
@@ -19,80 +19,71 @@ class Admin_Model_EventsMapper extends Admin_Model_DataMapper_Abstract
      * @var Admin_Model_DbTable_Events
      */
     protected $_dbTableClass = 'Admin_Model_DbTable_Events';
-    protected function createFromArray(array $array) {
-        ;
+    protected function createFromArray(array $array)
+    {
+        return new Admin_Model_Events($array);
     }
     public function save(Admin_Model_Events $Events)
     {
         $data = array(
             'title'                 => $Events->getTitle(),
-            'short_description'     => $Events->getShortDescription(),
-            'detail_description'    => $Events->getDetailDescription(),
-            'img_preview'           => $Events->getImgPreview(),
+            'short_description'     => $Events->getShort_description(),
+            'detail_description'    => $Events->getDetail_description(),
+            'img_preview'           => $Events->getImg_preview(),
             'published'             => $Events->getPublished(),
-            'created_at'            => $Events->getCreatedAt(),
-            'updated_at'            => $Events->getUpdatedAt(),
-            'published_at'          => $Events->getPublishedAt()
+            'updated_at'            => $Events->getUpdated_at(),
+            'published_at'          => $Events->getPublished_at()
         );
 
         if (null === ($id = $Events->getId())) {
+            $data['created_at'] = date("Y-m-d H:i:s");
             $this->getDbTable()->insert($data);
             $insertedId = $this->getDbTable()->getDefaultAdapter()->lastInsertId();
             $Events->setId($insertedId);
         } else {
             $this->getDbTable()->update($data, array('id = ?' => $id));
         }
-        return $Events;
     }
-    public function find($id, Admin_Model_Events $Events)
+    public function findById($id)
     {
         $result = $this->getDbTable()->find($id);
         if (0 == count($result)) {
-            throw new Zend_Exception("Events by id " . $id . " not found");
+            throw new UnexpectedValueException("Events by id " . $id . " not found");
         }
+        $object = new Admin_Model_Events();
         $row = $result->current();
-        $Events->setId($row->id)
-                  ->setTitle($row->title)
-                  ->setShortDescription($row->short_description)
-                  ->setDetailDescription($row->detail_description)
-                  ->setImgPreview($row->img_preview)
-                  ->setPublished($row->published)
-                  ->setCreatedAt($row->created_at)
-                  ->setUpdatedAt($row->updated_at)
-                  ->setPublishedAt($row->published_at);
+        $object->setOptions($row->toArray());
+        return $object;
     }
     /**
      * @param string|array|Zend_Db_Table_Select $where  OPTIONAL An SQL WHERE clause or Zend_Db_Table_Select object.
      * @param string|array                      $order  OPTIONAL An SQL ORDER clause.
      * @param int                               $count  OPTIONAL An SQL LIMIT count.
      * @param int                               $offset OPTIONAL An SQL LIMIT offset.
-     * @return Admin_Model_Events all set
+     * @return Admin_Model_EventsCollection all set
      */
     public function fetchAll($where = null, $order = null, $count = null, $offset = null)
     {
         $resultSet = $this->getDbTable()->fetchAll($where, $order, $count, $offset);
-        $entries   = array();
-        foreach ($resultSet as $row) {
-            $entry = new Admin_Model_Events();
-            $entry->setId($row->id)
-                  ->setTitle($row->title)
-                  ->setShortDescription($row->short_description)
-                  ->setDetailDescription($row->detail_description)
-                  ->setImgPreview($row->img_preview)
-                  ->setPublished($row->published)
-                  ->setCreatedAt($row->created_at)
-                  ->setUpdatedAt($row->updated_at)
-                  ->setPublishedAt($row->published_at)
-                  ->setMapper($this);
-            $entries[] = $entry;
-        }
-        return $entries;
+        return new Admin_Model_EventsCollection($resultSet->toArray(), $this);
     }
 
     public function delete($id)
     {
-        $where = $this->getDbTable()->getDefaultAdapter()->quoteInto('id = ?', $id);
-        $this->getDbTable()->delete($where);
+        $object = $this->findById($id);
+        if(null !== $object) {
+            $where = $this->getDbTable()->getDefaultAdapter()->quoteInto('id = ?', $id);
+            $this->getDbTable()->delete($where);
+        }
+    }
+
+    public function deleteBulk(array $ids)
+    {
+        if(count($ids) > 0) {
+            foreach($ids as $id) {
+                $this->delete($id);
+            }
+        }
     }
     /**
      * Sets published in where clause
@@ -102,8 +93,15 @@ class Admin_Model_EventsMapper extends Admin_Model_DataMapper_Abstract
      */
     public function published($published)
     {
-        $this->_select = $this->getSelect();
-        $this->_select->where('published = ?', $published);
+        /**
+         * isset added to prevent the clause when user is updated
+         * and $enabled value comes as null
+         */
+        if($published != '*' && isset($published)) {
+            $this->_select = $this->getSelect();
+            $this->_select->where('published = ?', $published);
+        }
+        return $this;
     }
     /**
      * Simple search by title field using like operator
@@ -113,7 +111,12 @@ class Admin_Model_EventsMapper extends Admin_Model_DataMapper_Abstract
      */
     public function search($value)
     {
-        $this->_select = $this->getSelect();
-        $this->_select->where('title LIKE ?', '%' . $value . '%');
+        if(!empty($value)) {
+            $value = str_replace('\\', '\\\\', $value);
+            $value = addcslashes($value, '_%');
+            $this->_select = $this->getSelect();
+            $this->_select->where('title LIKE ?', '%' . $value . '%');
+        }
+        return $this;
     }
 }
